moved gateway listen server to utils
