+++ server/src/main.rs
@@ -2,13 +2,10 @@ use std::env;
 use env_logger;
 use log::debug;
 
-use tokio::io::{AsyncReadExt, AsyncWriteExt};
-use tokio::net::{TcpListener, TcpStream};
+use tokio::net::TcpListener;
 use tokio::sync::{watch, broadcast};
 
 use watertank_simulation_server::utils::watertank::WaterTank;
-use watertank_simulation_server::utils::protocol::{Payload, ReturnMessage};
-use watertank_simulation_server::utils::protocol;
 
 use watertank_simulation_server::utils::simulation;
 
@@ -56,7 +53,7 @@ async fn main() {
     let (txin, rxin) = broadcast::channel(2);
 
     // Start and run the listeners and simulation.
-    let t = listen_tcp(gw_listener, rxout.clone(), txin.clone());
+    let t = server::listen_tcp(gw_listener, rxout.clone(), txin.clone());
     let ws = server::listen_ws(ws_listener);
     let r = simulation::run_simulation(txout, rxin, tank);
     r.await;
@@ -68,57 +65,4 @@ async fn main() {
     loop {
         thread::sleep(delay);
     }
-}
-
-async fn listen_tcp(listener: TcpListener, rxout: watch::Receiver<WaterTank>, txin: broadcast::Sender<Payload>) {
-    tokio::spawn(async move {
-        while let Ok((stream, addr)) = listener.accept().await {
-            debug!("New connection from {:?}", addr);
-            handle_gw(stream, rxout.clone(), txin.clone()).await;
-        }
-    });
-}
-
-async fn handle_gw(mut stream: TcpStream, rxout: watch::Receiver<WaterTank>, txin: broadcast::Sender<Payload>) {
-    
-    tokio::spawn(async move {
-        debug!("Handle new connection");
-
-        // In a loop, read data from the socket and write the data back.
-        loop {
-            let tank = *rxout.borrow();
-            let tank_level = protocol::convert_f32_to_mobdus_u16(0.0, tank.height, tank.level);
-            let tank_inflow = protocol::convert_f32_to_mobdus_u16(0.0, tank.max_inflow, tank.inflow);
-
-            let (mut reader, mut writer) = stream.split();
-
-            // read header length
-            let mut len = vec![0; 1];
-            match reader.peek(&mut len).await.unwrap() {
-                0 => {break},
-                _ => {},
-            };
-
-            let header = protocol::read_header(len, &mut reader).await;
-
-            // read payload
-            let payload = protocol::read_payload(header, &mut reader).await;
-            
-            debug!("Payload {:?}", payload);
-
-            txin.send(payload).unwrap();
-
-            // write something random
-            let hardcoded = ReturnMessage {
-                msg_type: String::from("input-register"),
-                address: 0,
-                tank_level: tank_level,
-                tank_inflow: tank_inflow,
-            };
-            let mut hardcoded = serde_json::to_string(&hardcoded).unwrap();
-            debug!("Sending {}", hardcoded);
-            hardcoded.push('\n');
-            writer.write_all(hardcoded.as_bytes()).await.unwrap();
-        }
-    });
 }
